[Harness] Move all project generation logic to the template object.
The template object is the one that knows how to create a new project,
move all the logic there and leave the generator as a class that simply
contains the definitions of knonw projects.

This means that when we move the code out, the generator will only be
used in the xamarin-macios repo, and other teams will use the template
and will pass the definition of the projects that they want to run, for
example, building the project definition from the command line
arguments.
