+++ tests/xharness/BCLTestImporter/BCLTestImportTargetFactory.cs
@@ -11,19 +11,15 @@ namespace Xharness.BCLTestImporter {
 	public class BCLTestImportTargetFactory {
 		public Harness Harness;
 		BCLTestProjectGenerator projectGenerator;
-		
+
 		public BCLTestImportTargetFactory (Harness harness)
 		{
 			Harness = harness;
 			var outputDir = Path.GetFullPath (Path.Combine (Harness.RootDirectory, "bcl-test"));
-			var projectTemplatePath = outputDir;
-			var registerTypesTemplatePath = Path.Combine (outputDir, "RegisterType.cs.in");
-			var plistTemplatePath = outputDir;
 
-			projectGenerator = new BCLTestProjectGenerator (outputDir, Harness.MONO_PATH, projectTemplatePath, registerTypesTemplatePath, plistTemplatePath) {
+			projectGenerator = new BCLTestProjectGenerator (outputDir, Harness.MONO_PATH) {
 				iOSMonoSDKPath = Harness.MONO_IOS_SDK_DESTDIR,
 				MacMonoSDKPath = Harness.MONO_MAC_SDK_DESTDIR,
-				Override = true,
 				GuidGenerator = Harness.NewStableGuid,
 				GroupTests = Harness.InCI || Harness.UseGroupedApps,
 			};
