+++ tests/xharness/BCLTestImporter/BclTestProjectInfo.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace Xharness.BCLTestImporter {
+	public class BclTestProjectInfo {
+		public string Name;
+		public string [] assemblies;
+		public string ExtraArgs;
+		public string Group;
+		public double TimeoutMultiplier = 1;
+	}
+}
