+++ tests/xharness/BCLTestImporter/Templates/ITemplatedProject.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 
@@ -7,11 +8,21 @@ namespace Xharness.BCLTestImporter.Templates {
 	// The interface should be able to generate a project that will later be
 	// used by the AppRunner to execute tests.
 	public interface ITemplatedProject {
-		Stream GetProjectTemplate (Platform platform);
-		Stream GetProjectTemplate (WatchAppType appType);
-		Stream GetPlistTemplate (Platform platform);
-		Stream GetPlistTemplate (WatchAppType appType);
-		Stream GetRegisterTypeTemplate ();
-		Task GenerateSource (string srcOuputPath);
+		string OutputDirectoryPath { get; set; }
+		string IgnoreFilesRootDirectory { get; set; }
+
+		IAssemblyLocator AssemblyLocator { get; set; }
+		IProjectFilter ProjectFilter { get; set; }
+		public Func<string, Guid> GuidGenerator { get; set; }
+
+		/// <summary>
+		/// Generates all the project files for the given projects and platform
+		/// </summary>
+		/// <param name="projects">The list of projects to be generated.</param>
+		/// <param name="platform">The platform to which the projects have to be generated. Each platform
+		/// has its own details.</param>
+		/// <param name="generatedDir">The dir where the projects will be saved.</param>
+		/// <returns></returns>
+		Task<List<BclTestProject>> GenerateTestProjectsAsync (IEnumerable<BclTestProjectInfo> projects, Platform platform);
 	}
 }
