+++ tests/xharness/xharness.csproj
@@ -152,6 +152,13 @@
     <Compile Include="XmlResultParser.cs" />
     <Compile Include="BCLTestImporter\Templates\ITemplatedProject.cs" />
     <Compile Include="BCLTestImporter\Templates\Managed\XamariniOSTemplate.cs" />
+    <Compile Include="BCLTestImporter\BclTestProjectInfo.cs" />
+    <Compile Include="BCLTestImporter\iOSBclTestProject.cs" />
+    <Compile Include="BCLTestImporter\BclTestProject.cs" />
+    <Compile Include="BCLTestImporter\Templates\IAssemblyLocator.cs" />
+    <Compile Include="BCLTestImporter\Templates\IProjectFilter.cs" />
+    <Compile Include="BCLTestImporter\Xamarin\AssemblyLocator.cs" />
+    <Compile Include="BCLTestImporter\Xamarin\ProjectFilter.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="..\..\tools\mtouch\SdkVersions.cs">
@@ -175,6 +182,7 @@
     <Folder Include="BCLTestImporter\Templates\Managed\Resources\src\common\TestRunner\Core\" />
     <Folder Include="BCLTestImporter\Templates\Managed\Resources\src\common\TestRunner\NUnit\" />
     <Folder Include="BCLTestImporter\Templates\Managed\Resources\src\common\TestRunner\xUnit\" />
+    <Folder Include="BCLTestImporter\Xamarin\" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="BCLTestImporter\Templates\Managed\Resources\Managed.iOS.csproj.in" />
